Upload generated sitemaps to configured S3 bucket.
Change the SitemapGenerator config to use an S3 bucket based on this wiki article:
https://github.com/kjvarga/sitemap_generator/wiki/Generate-Sitemaps-on-read-only-filesystems-like-Heroku
Add variable placeholders for S3 credentials / config.
Add route to serve sitemap from S3 bucket.
Tweak release-task to work on staging and production.
This may cause issues on review apps - will have to look / figure a strategy for this?