358: Fix issue with removing achievements.
When an Achievement had transitions associated with it, trying to remove this failed.
This seems to occur when achievements are created by importing from Future Learn - the online
courses are still self-certifiable, so a user may have tried to remove this, but failed.
Added a migration to fix the foreign keys on AchievementTransition and a test covering the error.