427: Add the omniauth-rails_csrf_protection and change auth request method
Omniauth has a possibly vulnerability that they are debating whether to fix or
leave 3rd parties to sort
From here: https://github.com/omniauth/omniauth/pull/809#issuecomment-497376674
Install the omniauth-rails_csrf_protection gem and tweak code for correct usage:
https://github.com/cookpad/omniauth-rails_csrf_protection#usage

The 'Log in' link now uses POST, but may also have a query param for a URI that
the user will be redirected to after a successful login.  This is potentially
frowned upon, but not detrimental - because of the way Omniauth works, it is
quite difficult to use the link_to / button_to helper to specify the source_uri param
as something in the POST request - any suggestions welcome.