419: WIP on 'calculating' the Certificate Number
We will count the number of 'passed' assessments to work out the number of the
certificate for a user.
Add a scope to the assessment_attempt model to return all passed assessments that
are sorted by date order.  Iterate through to find the 'index' of the current user.
This is our number.
POC working at 'runtime' by retrieving in the controller and using to display certificate
Add helper to render as per specs.

Issues:
1. In the ProgrammesController we are already getting the achievements (which we might also
use to calculate the number from the 'completed' assessment achievement transitions).
Consider refactoring to use this.
2. Calculating dynamically may mean the index could change if we remove a user (this cascades
a delete of their assessment transitions...): Consider storing the number as meta-data on their
transition to 'passed' for the assessment_attempt (see the UpdateUserAssessmentAttemptFromClassMarkerJob)
3. Consider moving the storing of this data in the Achievement transition (we'd maybe have to refactor
set_to_complete to allow optional meta-data for the transition).
4. Generally some of the controller stuff around certificates / assessments may need looking at so we can
streamline it, or be consistent, e.g. use Assessment for the exam only and Achievement for general stuff?