419: Refactor to set of the certificate_index on Achievement
Set the certificate_index on the Achievement when this is transitioned to 'passed'
by the job which updates this along with the user's AssessmentAttempt.
Allow Achievement.set_to_complete method to receive a hash of data that
it will set as the metadata on the transition.
Refactor ProgrammeController to obtain the certificate index from the AchievementTransition
Add a method to AchievementTransition to fix those missing metadata
Allow ProgrammeController to silently fix the transition for users who have already completed
the CS Accelerator.
Tests for above.
Delegate the 'last_transition' method for the *Transition models to the StateMachine so we can
call the method on the model instance.