419: Add counter model for tracking certificate numbers.
So we can have a consistent, non-breakable, numbering system for the certificates,
add an AssessmentCounter model that keeps track of the numbering.
This is updated when the job (triggered by ClassMarker's webhook) sets the assessment / achievement
to 'passed' for the corresponding user and the number is stored with their achievement.
If the user doesn't have a number, we will default to 0 in the view - we need to manually fix pre-existing
graduates with a rake task, etc.
Make naming a bit more consistent - number not index anymore.
Add tests where appropriate.
Add seed for DB