609: Create Presenters to help with varying achievement views
Create a number of Presenter classes with a common API so we can use these
to help render our 'Achievement Activities' in the user's programme 'Progress'
Each type of activity has a presenter that returns the varying content we use in
our view for an item in a list.
http://eewang.github.io/blog/2013/09/26/presenting-the-rails-presenter-pattern/

Factor the view partials to a single generic achievements template that is passed
an array of presenters to help render each achievement.

Where a user hasn't achieved the gaol, the Presenter is initialised with a nil object, but
it's declared methods can still be called, so we can generate the 'incomplete' activities.

Split out the variation from the Programmes controller by creating a 'Service'
type class to do the work of getting the Achievements for the specific programme.
http://sporto.github.io/blog/2012/11/15/a-pattern-for-service-objects-in-rails/

Change show & complete methods in controller to use the Service.
Remove unused templates.
Fix tests for generic template.

TODO:
- Tests for presenters
- Factor out the assessment variation from the controller
- Maybe create a Service for each different programme (perhaps with common ancestry?)
- Fix request specs for controller - stub lots of things.
- Fix view specs for show / complete