609: Factor out assessment details and move the 'services' to the models folder
Create a class to handle setting up and accessing the data that deals with
assessment attempts - use as a "View Model" kind of pattern.
Move the ProgrammeAchievementPresenters to UserProgrammeAchievements so it
is inline with the View Model idea and add accessors for attrs along with
an arg to control the number of items we want - test.
Add helper / scope to AssessmentAttempts model & test.
Fix tests