627: Refactor user_completed? to deal with programme variations
Using some of the ideas in here:
https://www.webascender.com/blog/tutorial-classes-inheritance-modules-mixins-ruby-rails/
Create modules for the CSAccelerator & PrimaryCertificate to handle the logic variation.
Delegate from the user_completed? method in Programme explicitly rather than trying to
be too clever.
Add tests for this and move out of Programmes spec.
Add scope to Achievement to filter by activity credit & test.