643: Add missing FutureLearn course and report missing ids to Raven
Throw an error when a FutureLearn course id is missing a matching Activity and
then send a message to Raven with details.  Try to only do this once for each
course id.
Add spec's to test this.
Comment out seeding of Primary & Secondary programmes along with incomplete
'community' activities temporarily until these are complete so we can
still run 'db:seed' in production without them for now.