Address 2 medium SQL injection risks highlighted by Brakeman
In Achiever, we use the transition @klass to get the table name - this shouldn't come from
outside Rails, but sanitize the SQL to remove risk.

In UserProgrammeAchievements we use a parameter passed into a method in the SQL.
The param comes from a view, but make sure this is escaped to remove risk.