801: Redirect to the user's last page on login.
From https://github.com/omniauth/omniauth/wiki/Saving-User-Location - use this
if we haven't specifically set a value.
Fix the overrides in the resources page to URL encode the #anchors.
Fix the #anchors in the resources templates.
Add #anchors to each course in the courses page and redirect to the specific anchor.
if we click on a course link when not logged in.
Fix heading structure in courses and secondary teachers page.