867: Add new states to the Achievement state-machine to track progress.
Change the 'commenced' state to 'enrolled' for clarity & fix spec's.
Add new states 'step_1' and 'step_2' to state-machine with allowed transitions with tests.
Add update_state method to Achievement to handle the logic of setting the right state
based on progress and whether user has left a course.
Refactor FetchUsersCompletedCoursesFromAchieverJob to use update_state
Update and fix spec's for FetchUsersCompletedCoursesFromAchieverJob and other fallout.