Ensure assessment_eligibility_job only requested once
When processing multiple courses the assessment ability job could be queued multiple times in rapid succession.
This was leading to a race condition where the record to say a user has already been sent an email was being created after the next job had checked that it didn't exist.

This was then causing an error in postgres as trying to insert the record would break unique validation.

Instead of queuing a job for each record in the csv or list of user's course only queue the job once per user.