Fix ProgrammeActivityGrouping achievement caching issue, rm includes
We found when debugging a rake task that called user_complete on the
same ProgrammeActivityGrouping that the instance variable was the same,
IE containing user information for other users, because it was the same
PAG instance. We've removed the IVAR memoization to fix this.

We also noticed that with the way its currently implemented, the
benefit that was trying to be derived by moving it into an IVAR was
being missed as it was still an active record collection and so
calling find_by was still doing a DB query. We've changed this to
use ruby's Array#find

We've also removed the excessive #includes on the user.achievements
query as it seems it was never used

Restore contract of ProgrammeActivityGrouping#achievements(user)

Rm N+1 - no need to load activity when we are only comparing activity_id