Removing old programme activities from primary.
## Now

I've made a modification to the modal component which allows you to
bring your own close button. I've also written tests for the modal
as I seem to have neglected to add any the first time round.

I've added the concept of "message_flags" on UPEs that can be used
to enable a message to be shown to a user. I've not added a generic
message flag controller and instead opted to just have a single
action on the UPE controller as I don't want to build too much of an
abstraction which we may not use.