Add PDL as I Belong course option
# Now

This commit includes the PDLP courses in the lists that are fetched
from dynamics. There are changes to the spec stubs in accordance to that.

The logic for the IB page is that we should only show the CPD course if
no courses have been completed yet. If either the CPD or PDL course have been
completed, we should show that, if both have been completed, we should show that.

To guarentee the CPD comes first, in the IB programme seed, I specify the order
and use .first(1) to only load the first in the frontend