Made achievements global
# Before

Before this commit, achievements were assigned to one particular
programme. We however have activities which are part of multiple
programmes. This was causing issues as the app didn't know how
to handle having multiple achievements assigned to the same user
for the same activity.

# Now

I considered removing the unique constraint and allowing multiple
complete achievements, but I instead decided to simplify some of
the old logic, and have achievements be "global" such that
completing an activity in Primary, would also cause it to be
completed in Secondary.