Added CSA to secondary computing dashboard.
# Now
This PR adds the CSA as an item in the progress bar at the top of the secondary programme dashboard.

We've added a handful of new programmes to T/C now and I've held off "drying" up the code as each of the programmes has had a quite distinct set of requirements for completion. In this PR, I've started an abstraction which should let us get rid of having multiple programme controllers, and move to having just one generic controller in the future.

It's clear that we have many different completion requirements for each of our programmes, some being the number of activities needing to be completed, others requiring other programmes to be completed first, and others requiring quizzes to be completed.

I've started on a new abstraction for "ProgrammeObjective"s. The ProgrammeObjective role has been designed to conform to part of the existing interface of the ProgrammeActivityGroupings. Primarily the "user_complete?(user)" method; thus with a little tweaking, making ProgrammeActivityGroupings a programme objective.

I've changed the old logic in the base Programme class which iterated over all the ProgrammeActivityGroupings, and now ask it to iterate over all the ProgrammeObjectives. For most Programmes the list of ProgrammeObjectives will the the same as the ProgrammeActivityGroupings, but it allows us to create different objectives that can test different actions the user has made.

ProgrammeObjectives also have a "objective_displayed_in?(location)" method. Currently, we have two places where we want to display the progress to users. The top "progress bar" and in the main "body" of the programme dashboards. This PR has modified the ProgressBarComponent to use the list of ProgrammeObjectives which are displayed in the :progress_bar, allowing in this case for the CSA programme completion to be one of the steps that needs completing.

In the future, we will be able to create a new generic programme controller which similarly iterates all the ProgrammeObjectives that belong to the body, rendering a ViewComponent which is provided via a method on the ProgrammeObjectives playing the body role, and dynamically render the relevant UI elements.