Introduced report generation
For anyone wanting to change the report generation logic or the programme
activity grouping completion checking logic, be very careful as those
are very hot and if they contain an N+1 query, it would be very bad
for the performance of report generation.

This commit contains a reasonable change to the PAG completion logic
and rewrites it in a way inspired by graphql dataloaders. This change
allows us to use the same logic in the report generation as well as
fufilling the previous interface