Remove line for non online courses to better match designs
Add a no_top_border option to ProviderLogosComponent to allow
removing the top border and padding when needed.

Update the _provider_logos_component.scss to include a modifier class
for no top border. Modify the ProviderLogosComponent to accept the
no_top_border option and set the corresponding class.

Use the new option in _booking_list.html.erb to conditionally remove
the top border when there is no current user and the course is not an
online CPD or has not started.