Updating components to match the new design for the TCC page.
Created side banner component that has been added to the two column picture section. It does basic structure, but will need to be positioned and adjusted in whatever component its imported into.

Modified the two column video component to allow the video the be placed at either side.

Added option to pass into the rich blocks to set the class on lists.

Fixing error in the mock template system